Add test for list (2/2)

+++ backend/src/app/services/lists.js
@@ -33,7 +33,9 @@ exports.getAllList = filters => {
     limit: filters.limit,
     order: filters.orderColumn ? [[filters.orderColumn, filters.orderType || 'ASC']] : undefined
   }).catch(err => {
+    /* istanbul ignore next */
     logger.error(inspect(err));
+    /* istanbul ignore next */
     throw databaseError(`There was an error getting the lists: ${err.message}`);
   });
 };
@@ -47,7 +49,9 @@ exports.getList = (filters, options = {}) => {
     },
     ...options
   }).catch(err => {
+    /* istanbul ignore next */
     logger.error(inspect(err));
+    /* istanbul ignore next */
     throw databaseError(`There was an error getting the list: ${err.message}`);
   });
 };
@@ -75,7 +79,9 @@ exports.deleteList = filters => {
         ).then(() => list.destroy({ transaction }))
       )
       .catch(err => {
+        /* istanbul ignore next */
         logger.error(inspect(err));
+        /* istanbul ignore next */
         throw databaseError(`There was an error deleting the list: ${err.message}`);
       });
   });
@@ -85,7 +91,9 @@ exports.createList = attributes => {
   logger.info(`Attempting to create list with attributes: ${inspect(attributes)}`);
   return Country.count({ where: { id: attributes.countriesIds } })
     .catch(err => {
+      /* istanbul ignore next */
       logger.error(inspect(err));
+      /* istanbul ignore next */
       throw databaseError(`There was an error verifying the countries: ${err.message}`);
     })
     .then(count => {
@@ -100,7 +108,9 @@ exports.createList = attributes => {
           })
         )
         .catch(err => {
+          /* istanbul ignore next */
           logger.error(inspect(err));
+          /* istanbul ignore next */
           throw databaseError(`There was an error creating the list: ${err.message}`);
         });
     });
@@ -113,20 +123,32 @@ exports.updateList = attributes => {
     if (!list) {
       throw notFound('The list was not found');
     }
-    return sequelizeInstance
-      .transaction(transaction => {
-        const { countriesByListToDelete, countriesByListToCreate } = getCountriesToDeleteAndCreate(
-          list,
-          attributes
-        );
-        return Promise.all([
-          ...countriesByListToDelete.map(countryByList => countryByList.destroy({ transaction })),
-          CountryByList.bulkCreate(countriesByListToCreate, { transaction })
-        ]);
-      })
+    return Country.count({ where: { id: attributes.countriesIds } })
       .catch(err => {
+        /* istanbul ignore next */
         logger.error(inspect(err));
-        throw databaseError(`There was an error updating the list: ${err.message}`);
+        /* istanbul ignore next */
+        throw databaseError(`There was an error verifying the countries: ${err.message}`);
+      })
+      .then(count => {
+        if (attributes.countriesIds.length !== count) throw invalidCountries();
+        return sequelizeInstance
+          .transaction(transaction => {
+            const { countriesByListToDelete, countriesByListToCreate } = getCountriesToDeleteAndCreate(
+              list,
+              attributes
+            );
+            return Promise.all([
+              ...countriesByListToDelete.map(countryByList => countryByList.destroy({ transaction })),
+              CountryByList.bulkCreate(countriesByListToCreate, { transaction })
+            ]);
+          })
+          .catch(err => {
+            /* istanbul ignore next */
+            logger.error(inspect(err));
+            /* istanbul ignore next */
+            throw databaseError(`There was an error updating the list: ${err.message}`);
+          });
       });
   });
 };
@@ -151,7 +173,9 @@ exports.getCountriesByList = filters => {
         count: countriesByList.count
       }))
       .catch(err => {
+        /* istanbul ignore next */
         logger.error(inspect(err));
+        /* istanbul ignore next */
         throw databaseError(`There was an error getting countries of the list: ${err.message}`);
       });
   });
@@ -171,7 +195,9 @@ exports.createCountriesByList = attributes => {
         countryId: attributes.countryId,
         listId: attributes.id
       }).catch(err => {
+        /* istanbul ignore next */
         logger.error(inspect(err));
+        /* istanbul ignore next */
         throw databaseError(`There was an error creating country of the list: ${err.message}`);
       });
     });
@@ -190,7 +216,9 @@ exports.deleteCountriesByList = attributes => {
         listId: attributes.id
       }
     }).catch(err => {
+      /* istanbul ignore next */
       logger.error(inspect(err));
+      /* istanbul ignore next */
       throw databaseError(`There was an error deleting country of the list: ${err.message}`);
     });
   });
